requirements.txt
cached-property==1.5.2
certifi==2020.11.8
chardet==3.0.4
cycler==0.10.0
Cython==0.29.21
dataclasses==0.8
DateTime==4.3
efficientnet-pytorch==0.7.0
future==0.18.2
h5py==3.1.0
idna==2.10
kiwisolver==1.3.1
matplotlib==3.3.3
numexpr==2.7.1
numpy==1.19.4
opencv-python==4.4.0.46
panda==0.3.1
pandas==1.1.4
Pillow==8.0.1
pycocotools==2.0.2
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.4
requests==2.25.0
six==1.15.0
tables==3.6.1
torch==1.7.0
torchsummary==1.5.1
torchvision==0.8.1
tqdm==4.52.0
typing-extensions==3.7.4.3
urllib3==1.26.2
zope.interface==5.2.0